refactor: Service 클래스 패키지 재배치
