+++ src/main/java/BE_Elixir/Elixir/domain/challenge/controller/ChallengeController.java
@@ -4,8 +4,8 @@ package BE_Elixir.Elixir.domain.challenge.controller;
 import BE_Elixir.Elixir.domain.challenge.controller.api.ChallengeApi;
 import BE_Elixir.Elixir.domain.challenge.dto.request.ChallengeRequestDTO;
 import BE_Elixir.Elixir.domain.challenge.dto.response.*;
-import BE_Elixir.Elixir.domain.challenge.dto.service.ChallengeAchievementService;
-import BE_Elixir.Elixir.domain.challenge.dto.service.ChallengeService;
+import BE_Elixir.Elixir.domain.challenge.service.ChallengeAchievementService;
+import BE_Elixir.Elixir.domain.challenge.service.ChallengeService;
 import BE_Elixir.Elixir.domain.member.entity.Member;
 import BE_Elixir.Elixir.domain.member.entity.MemberDetails;
 import BE_Elixir.Elixir.global.response.CommonResponse;
