+++ src/main/java/BE_Elixir/Elixir/domain/challenge/event/listener/ChallengeEventListener.java
@@ -6,7 +6,7 @@ import BE_Elixir.Elixir.domain.challenge.event.events.DietLogEvent;
 import BE_Elixir.Elixir.domain.challenge.event.events.RecipeEvent;
 import BE_Elixir.Elixir.domain.challenge.repository.ChallengeAchievementRepository;
 import BE_Elixir.Elixir.domain.challenge.repository.ChallengeRepository;
-import BE_Elixir.Elixir.domain.challenge.dto.service.ChallengeAchievementService;
+import BE_Elixir.Elixir.domain.challenge.service.ChallengeAchievementService;
 import BE_Elixir.Elixir.domain.dietLog.repository.DietLogRepository;
 import BE_Elixir.Elixir.domain.ingredient.entity.Ingredient;
 import BE_Elixir.Elixir.domain.ingredient.repository.IngredientRepository;
